requirements.txt
click>=8.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
tabulate>=0.9.0
colorama>=0.4.4
diskcache>=5.4.0
prompt-toolkit>=3.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
requests>=2.25.0
google-genai>=0.1.0
